fix(ads): DBFCDX conformance sweep + pre-existing test failures
Found by the new dballcmp/navcmp Harbour harnesses (DBFCDX golden
reference, transcripts byte-compared), reported by Pritpal Bedi:

- AdsGetLastTableUpdate: copy_ace_string reserved a NUL byte, so the
  8-digit date truncated to 7 chars when rddads passed its leftover
  AdsGetDateFormat length as capacity (LUpdate() showed empty).
- AdsGetIndexFilename ignored ADS_FULLPATHNAME/BASENAMEANDEXT/BASENAME
  (OrdBagName returned a full path instead of the bare bag name).
- index_expr evaluator had no arithmetic level: OrdCondSet+OrdCreate
  with FOR "NUM % 2 == 0" parsed as bare NUM (truthy) and indexed the
  wrong rows. Added + - * / % between atoms.
- REINDEX/OrdListRebuild dropped deleted rows' keys (CREATE INDEX kept
  them); DBFCDX keeps them in both. Fixes OrdKeyCount drift.
- AdsGetRelKeyPos now uses the ADS bucket-centre convention
  ((n + 0.5) / count) rddads documents.
- AdsDeleteIndex only detached the tag; it now rewrites the CDX bag
  via the new CdxIndex::delete_tag, so OrdDestroy persists.
- Production-bag semantics: structural-bag tags stay attached through
  OrdListClear/OrdBagClear while the table is open (DBFCDX parity);
  OrdDestroy still removes a structural tag permanently.
- AdsFindFirstTable/AdsFindNextTable returned AE_NO_MATCHING_FILE
  (5059) instead of AE_NO_FILE_FOUND (5018) that rddads loops on.
- Remote AdsDropTable: the in-process server's ABI-twin call took the
  any-remote fallback and hijacked the client's wire connection
  (recv() failed). Same local-handle guard as AdsCreateTable.
- CI: -Wsign-conversion in tools/serverd/config_ini.cpp broke the
  Linux/macOS release legs.
- tests: ttest.cmd selective runner (pattern/all/slow/list); navcmp
  gains the SIXDRIVER-style OrdScope battery; dballcmp normalises the
  rddads-side gaps (OrdFindRec/OrdSkipRaw/OrdWildSeek unmapped,
  OrdKeyAdd/Del superset); sql_temp_browse_nav relpos expectations
  updated to the ADS convention.

+++ tools/serverd/config_ini.cpp
@@ -102,7 +102,7 @@ bool parse_ini(const std::string& text, IniConfig& out, std::string& error) {
         // Inside a [port:NNNN] section — accept data= only
         if (current_port_idx >= 0) {
             if (key == "data" || key == "data_dir" || key == "datadir") {
-                out.extra_ports[current_port_idx].data_dir = val;
+                out.extra_ports[static_cast<std::size_t>(current_port_idx)].data_dir = val;
             } else {
                 error = "line " + std::to_string(lineno) +
                         ": unknown key '" + key + "' in port section";
